Criada e finalizada tela AboutColabuco, adicionada no AppRoutes. Movidos os componentes da Home na pasta home. Criadas instâncias do componente Footer nas telas principais
