+++ src/components/Footer.jsx
@@ -3,8 +3,8 @@ import { Box } from '@chakra-ui/react';
 function Footer() {
     return (
         <Box mt='140px'>
-            <Box bg='#D9D9D9' h='430px'></Box>
-            <Box bg='#9E9E9E' h='330px'></Box>
+            <Box bg='#FFFF00' h='430px'></Box>
+            <Box bg='#6600CC' h='330px'></Box>
         </Box>
     );
 }
