+++ src/components/Navbar.jsx
@@ -47,7 +47,7 @@ function Navbar() {
                 <Link borderBottom='solid 3px transparent' style={{ textDecoration: 'none' }} href="/" className='fonte2 categoria'>INÍCIO</Link>
                 <Link borderBottom='solid 3px transparent' style={{ textDecoration: 'none' }} href="productscreen" className='fonte2 categoria'>PRODUTOS</Link>
                 <Link borderBottom='solid 3px transparent' style={{ textDecoration: 'none' }} href="" className='fonte2 categoria'>ARTISTAS</Link>
-                <Link borderBottom='solid 3px transparent' style={{ textDecoration: 'none' }} href="" className='fonte2 categoria'>SOBRE A COLABUCO</Link>
+                <Link borderBottom='solid 3px transparent' style={{ textDecoration: 'none' }} href="sobrecolabuco" className='fonte2 categoria'>SOBRE A COLABUCO</Link>
             </HStack>
         </Box>
     );
