+++ src/components/home/BuyByArtist.jsx
@@ -1,10 +1,10 @@
 import { Box, Center } from '@chakra-ui/react';
-import SectionTitle from "./SectionTitle";
-import ArtistProfile from './ArtistProfile';
-import artist1 from '../assets/artist1.png';
-import artist2 from '../assets/artist2.png';
-import artist3 from '../assets/artist3.png';
-import artist4 from '../assets/artist4.png';
+import SectionTitle from "../SectionTitle";
+import ArtistProfile from '../ArtistProfile';
+import artist1 from '../../assets/artist1.png';
+import artist2 from '../../assets/artist2.png';
+import artist3 from '../../assets/artist3.png';
+import artist4 from '../../assets/artist4.png';
 
 
 function BuyByArtist() {
