+++ src/components/home/TopSelling.jsx
@@ -1,14 +1,11 @@
 import { Center, Box } from '@chakra-ui/react';
-import CardProduct from './CardProduct';
-import '../styles/home.css'
-import SectionTitle from './SectionTitle';
-import ProductFilters from './ProductFilters';
+import CardProduct from '../CardProduct';
+import SectionTitle from '../SectionTitle';
 
-function Offers(props) {
+function TopSelling() {
     return (
         <Box my={12}>
-            <ProductFilters></ProductFilters>
-            <SectionTitle title='Ofertas'/>
+            <SectionTitle title='Produtos mais vendidos' />
             <Center>
                 <CardProduct />
                 <CardProduct />
@@ -20,4 +17,4 @@ function Offers(props) {
     );
 }
 
-export default Offers;
+export default TopSelling;
